# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HanapCare Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wards:
    get:
      operationId: listWards
      tags: [wards]
      summary: List wards
      responses:
        "200":
          description: Ward list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ward"
    post:
      operationId: createWard
      tags: [wards]
      summary: Create ward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardInput"
      responses:
        "201":
          description: Ward created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ward"
  /beds:
    get:
      operationId: listBeds
      tags: [wards]
      summary: List beds
      parameters:
        - name: wardId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bed list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bed"
    post:
      operationId: createBed
      tags: [wards]
      summary: Create bed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BedInput"
      responses:
        "201":
          description: Bed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bed"
  /beds/{id}:
    patch:
      operationId: updateBed
      tags: [wards]
      summary: Update bed status or assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BedUpdate"
      responses:
        "200":
          description: Updated bed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bed"

  # ─── STAFF ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Ward:
      type: object
      required: [id, name, wardType, totalBeds]
      properties:
        id:
          type: integer
        name:
          type: string
        wardType:
          type: string
        floor:
          type: ["string", "null"]
        totalBeds:
          type: integer
        availableBeds:
          type: integer
        description:
          type: ["string", "null"]
    WardInput:
      type: object
      required: [name, wardType, totalBeds]
      properties:
        name:
          type: string
        wardType:
          type: string
        floor:
          type: string
        totalBeds:
          type: integer
        description:
          type: string
    Bed:
      type: object
      required: [id, wardId, bedNumber, status]
      properties:
        id:
          type: integer
        wardId:
          type: integer
        wardName:
          type: ["string", "null"]
        bedNumber:
          type: string
        status:
          type: string
        patientId:
          type: ["integer", "null"]
        patientName:
          type: ["string", "null"]
        admittedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    BedInput:
      type: object
      required: [wardId, bedNumber]
      properties:
        wardId:
          type: integer
        bedNumber:
          type: string
        notes:
          type: string
    BedUpdate:
      type: object
      properties:
        status:
          type: string
        patientId:
          type: ["integer", "null"]
        notes:
          type: string

    # ─── STAFF ─────────────────────────────────────────────────
...